requirements.txt
# pip install -r requirements.txt (instalar as bibliotecas)
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dataclasses-json==0.6.7
deepdiff==8.0.1
Deprecated==1.2.14
emoji==2.14.0
filetype==1.2.0
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.3.2
langchain-community==0.3.1
langchain-core==0.3.8
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.131
lxml==5.3.0
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
orderly-set==5.2.2
orjson==3.10.7
packaging==24.1
pdf2image==1.17.0
pdfminer.six==20240706
pikepdf==9.3.0
pillow==10.4.0
psutil==6.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pypdf==5.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-oxmsg==0.0.1
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.10.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
tabulate==0.9.0
tenacity==8.5.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
unstructured==0.15.13
unstructured-client==0.25.9
urllib3==2.2.3
wrapt==1.16.0
yarl==1.13.1
